pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.1"
numpy = "^1.24.3"
torch = "2.0.1"
tqdm = "^4.65.0"
scipy = "^1.10.1"
librosa = "^0.10.0.post2"
torchshow = "^0.5.0"
torchaudio = "2.0.2"
llvmlite = "^0.40.1"
setuptools = "^68.0.0"
tensorboard = "^2.13.0"
torchvision = "^0.15.2"
pandas = "^2.0.3"
ipython = "^8.14.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
